Add geigen-based method for the LDA.
Also change test_lda.py to take into account that eigenvectors may be
parallel but not equal (but still unitary).

FIXME: this is currently hacky, in the following places:
- integration of the mode argument to choose the solver (if kept, should
be string, documented and tested)
- definition of the solver trait and impls should be in own file instead
of lda.rs
- eigenvectors selection should simply take the last p eigenvectors (as
both LAPACK and Eigen sort eigenvalues by ascending value (or
magnitude, I don't remember but it doesn't matter here since they are
positive).